us-core-r4/src/test/java/gov/va/api/health/r4/api/datatypes/TriggerDefinitionTest.java
package gov.va.api.health.r4.api.datatypes;

import static org.assertj.core.api.Assertions.assertThat;

import gov.va.api.health.r4.api.elements.Reference;
import java.util.Set;
import javax.validation.ConstraintViolation;
import javax.validation.Validation;
import javax.validation.ValidatorFactory;
import org.junit.jupiter.api.Test;

public class TriggerDefinitionTest {
  @Test
  void validationFailsGivenMultipleValues() {
    var bad =
        TriggerDefinition.builder()
            .type(TriggerDefinition.Type.data_accessed)
            .timingTiming(Timing.builder().build())
            .timingReference(Reference.builder().build())
            .build();
    assertThat(violationsOf(bad)).isNotEmpty();
  }

  @Test
  void validationPassesGivenGoodDefinition() {
    var good = TriggerDefinition.builder().type(TriggerDefinition.Type.data_accessed).build();
    assertThat(violationsOf(good)).isEmpty();
    good.timingTiming(Timing.builder().build());
    assertThat(violationsOf(good)).isEmpty();
  }

  private <T> Set<ConstraintViolation<T>> violationsOf(T object) {
    ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
    return factory.getValidator().validate(object);
  }
}
